+++ NEWS.md
@@ -11,7 +11,8 @@ New models supported:
 
 New features:
 
-* The new `averages()` function is the recommended way to "average over" (marginalize) unit-level estimates for all functions. It is more efficient, flexible, and explicit that `summary()` or `tidy()`.
+* New functions: `avg_predictions()`, `avg_comparisons()`, `avg_slopes()`
+* Equivalence, non-inferiority, and non-superiority tests with the `hypotheses()` function and `equivalence` argument.
 * New `df` argument to set degrees of freedom manually for p and CI.
 * `by` argument
   - `TRUE` returns average (marginal) predictions, comparisons, or slopes.